requirements.txt
asgiref==3.8.1
asn1crypto==1.5.1
blinker==1.7.0
boto3==1.34.92
botocore==1.34.92
Brlapi==0.8.5
certifi==2023.11.17
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==41.0.7
cupshelpers==1.0
dbus-python==1.3.2
defusedxml==0.7.1
diff-match-patch==20230430
distro==1.9.0
distro-info==1.7
Django==5.0.4
django-allauth==0.62.1
django-cors-headers==4.3.1
django-import-export==3.3.8
django-storages==1.14.2
dlib==19.24.4
et-xmlfile==1.1.0
face-recognition==1.3.0
face_recognition_models==0.3.0
html5lib==1.1
httplib2==0.20.4
idna==3.6
jmespath==1.0.1
kali-tweaks==2023.3.2
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
louis==3.29.0
MarkupPy==1.14
mechanize==0.4.9
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
pillow==10.3.0
pycairo==1.26.0
pycups==2.0.1
PyGObject==3.47.0.dev0
PyJWT==2.7.0
pyparsing==3.1.1
pysmbc==1.0.25.1
python-apt==2.7.6
python-dateutil==2.9.0
pyxdg==0.28
PyYAML==6.0.1
requests==2.31.0
s3transfer==0.10.1
six==1.16.0
sqlparse==0.5.0
tablib==3.5.0
urllib3==1.26.18
wadllib==1.3.6
webencodings==0.5.1
xdg==5
xlrd==2.0.1
xlwt==1.3.0
